[RENAME] 패키지 네이밍 규칙에 따라 스네이크_케이스로 변경 후 이름변경한 패키지들은 단수를 사용함
