+++ src/main/java/com/server/EZY/model/member/MemberEntity.java
@@ -1,7 +1,7 @@
 package com.server.EZY.model.member;
 
 import com.server.EZY.model.baseTime.BaseTimeEntity;
-import com.server.EZY.model.member.enumType.Role;
+import com.server.EZY.model.member.enum_type.Role;
 import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
 import com.fasterxml.jackson.annotation.JsonProperty;
