+++ src/main/java/com/server/EZY/model/plan/embedded_type/Period.java
@@ -1,4 +1,4 @@
-package com.server.EZY.model.plan.embeddedTypes;
+package com.server.EZY.model.plan.embedded_type;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
