+++ src/main/java/com/server/EZY/model/plan/errand/dto/ErrandSetDto.java
@@ -1,8 +1,8 @@
 package com.server.EZY.model.plan.errand.dto;
 
 import com.server.EZY.model.member.MemberEntity;
-import com.server.EZY.model.plan.embeddedTypes.Period;
-import com.server.EZY.model.plan.embeddedTypes.PlanInfo;
+import com.server.EZY.model.plan.embedded_type.Period;
+import com.server.EZY.model.plan.embedded_type.PlanInfo;
 import com.server.EZY.model.plan.errand.ErrandEntity;
 import com.server.EZY.model.plan.errand.ErrandStatusEntity;
 import lombok.*;
