+++ src/main/java/com/server/EZY/model/plan/errand/enum_type/ResponseStatus.java
@@ -1,4 +1,4 @@
-package com.server.EZY.model.plan.errand.enumType;
+package com.server.EZY.model.plan.errand.enum_type;
 
 public enum ResponseStatus {
     CANCEL, ACCEPT, NOT_READ, READ
