+++ src/main/java/com/server/EZY/model/plan/personal/PersonalPlanEntity.java
@@ -1,9 +1,8 @@
 package com.server.EZY.model.plan.personal;
 
-import com.server.EZY.exception.user.exception.InvalidAccessException;
 import com.server.EZY.model.plan.PlanEntity;
-import com.server.EZY.model.plan.embeddedTypes.Period;
-import com.server.EZY.model.plan.embeddedTypes.PlanInfo;
+import com.server.EZY.model.plan.embedded_type.Period;
+import com.server.EZY.model.plan.embedded_type.PlanInfo;
 import com.server.EZY.model.member.MemberEntity;
 import com.server.EZY.model.plan.tag.TagEntity;
 import lombok.*;
