+++ src/main/java/com/server/EZY/model/plan/tag/TagEntity.java
@@ -1,7 +1,7 @@
 package com.server.EZY.model.plan.tag;
 
 import com.server.EZY.model.member.MemberEntity;
-import com.server.EZY.model.plan.tag.embeddedTypes.Color;
+import com.server.EZY.model.plan.tag.embedded_type.Color;
 import lombok.*;
 
 import javax.persistence.*;
