+++ src/main/java/com/server/EZY/model/plan/tag/dto/TagSetDto.java
@@ -2,7 +2,7 @@ package com.server.EZY.model.plan.tag.dto;
 
 import com.server.EZY.model.member.MemberEntity;
 import com.server.EZY.model.plan.tag.TagEntity;
-import com.server.EZY.model.plan.tag.embeddedTypes.Color;
+import com.server.EZY.model.plan.tag.embedded_type.Color;
 import lombok.*;
 
 import javax.validation.constraints.NotNull;
