+++ src/test/java/com/server/EZY/model/member/MemberEntityTest.java
@@ -1,6 +1,6 @@
 package com.server.EZY.model.member;
 
-import com.server.EZY.model.member.enumType.Role;
+import com.server.EZY.model.member.enum_type.Role;
 import com.server.EZY.model.member.repository.MemberRepository;
 import com.server.EZY.testConfig.QueryDslTestConfig;
 import org.junit.jupiter.api.DisplayName;
